+++ packages/@enotion/config/index.d.ts
@@ -1,0 +1,9 @@
+declare module "@enotion/config";
+declare module "@enotion/config/constants";
+declare module "@enotion/config/tsup/tsup.config";
+declare module "@enotion/config/tsup/tsup.entry.config";
+declare module "@enotion/config/jest/jest.config";
+declare module "@enotion/config/jest/jest.server.config";
+declare module "@enotion/config/typescript/react-library";
+declare module "@enotion/config/typescript/nextjs";
+declare module "@enotion/config/typescript/base";