+++ CHANGELOG.md
@@ -26,6 +26,9 @@ All notable changes to this project will be documented in this file.
 - (**changed**) Moved typescript-config into a config package (@enotion/config/typescript)
 - (**changed**) Moved jest-config into a config package (@enotion/config/jest)
 - (**changed**) Added tsup-config into a config package (@enotion/config/tsup)
+- (**added**) Added constants into a config package (@enotion/config/constants)
+- (**changed**) Updated package.json exports to reflect neater import paths
+- (**changed**) Updated TypeScript `base` config to include ESNext in library target
 
 ##### Hooks